Revert "Upstream changes. (#2)"
This reverts commit e222eddfc448c6c40f17f45d497e765d753a1852.

+++ ExtLibs/Grid/GridUI.cs
@@ -29,8 +29,8 @@ namespace MissionPlanner
         private static readonly ILog log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
 
         // Variables
-        const double rad2deg = (180 / Math.PI);
-        const double deg2rad = (1.0 / rad2deg);
+        const float rad2deg = (float)(180 / Math.PI);
+        const float deg2rad = (float)(1.0 / rad2deg);
 
         private GridPlugin plugin;
         static public Object thisLock = new Object();
